law/requirements.txt
openai>=1.0.0
langchain>=0.1.0
faiss-cpu>=1.7.4
numpy>=1.21.0
tqdm>=4.62.0